refactor: separate candidates views into splited files
